fix(Actor): change data type long id to Integer id

+++ src/main/java/Task6/HerokunateAPI/Models/Actor.java
@@ -30,7 +30,7 @@ public class Actor {
     @Column
     private String url;
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
